+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `JewishCalendar#taanis_bechorim?` detects Taanis Bechorim for a given date
+
 ## [0.4.0] - 2020-10-13
 ### Added
 - `JewishCalendar#end_of_week` returns Shabbos for this week